requirements.txt
flask
flask-socketio
flask-mail
python-dotenv
email-validator
```